real result logs

+++ scene_predictor/normalize_data.py
@@ -11,6 +11,42 @@ import multiprocessing as mp
 import shutil
         
 
+def calculate_box_corners(x, y, theta, w, h):
+    """
+    Calculate the bottom-left and top-right corners of a 2D box given its center,
+    width, height, and rotation angle.
+    
+    Parameters:
+    - x, y: Coordinates of the box's center.
+    - theta: Rotation angle of the box in radians.
+    - w, h: Width and height of the box.
+    
+    Returns:
+    - (bl_x, bl_y, tr_x, tr_y): Coordinates of the bottom-left and top-right corners.
+    """
+    def rotate_point(px, py, theta):
+        """Rotate a point by theta around the origin."""
+        cos_theta, sin_theta = np.cos(theta), np.sin(theta)
+        x_rotated = px * cos_theta - py * sin_theta
+        y_rotated = px * sin_theta + py * cos_theta
+        return x_rotated, y_rotated
+
+    # Half dimensions
+    half_w, half_h = w / 2, h / 2
+    
+    # Calculate offsets for corners in local box coordinates
+    offsets = [(-half_w, -half_h), (half_w, half_h)]
+    
+    # Rotate offsets and translate to world coordinates
+    corners = [rotate_point(ox, oy, theta) for ox, oy in offsets]
+    corners_world = [(x + cx, y + cy) for cx, cy in corners]
+    
+    # Bottom-left and top-right corners
+    bl_x, bl_y = corners_world[0]
+    tr_x, tr_y = corners_world[1]
+    
+    return bl_x, bl_y, tr_x, tr_y
+
 def balance_data(files, dest_path):
     ctr = 0
     data_ctr = {
@@ -52,6 +88,26 @@ def balance_data(files, dest_path):
             data = np.load(file)
             last_idx = data['done'].nonzero()[0][-1]
             target = data['target_env']
+
+            # np.zeros(((last_idx, 8)))
+            # x = target[100, 2]
+            # y = target[100, 3]
+            # theta = target[100, 4]
+            # w = target[100, 5]
+            # h = target[100, 6]
+
+            # print(x, y, theta, w, h)
+            # print(calculate_box_corners(x, y, theta, w, h))
+            
+            # x = target[100, 11]
+            # y = target[100, 12]
+            # theta = target[100, 13]
+            # w = target[100, 14]
+            # h = target[100, 15]
+
+            # print(x, y, theta, w, h)
+            # print(calculate_box_corners(x, y, theta, w, h))
+
             index_found = int(int(target[last_idx, 2] != 0) + int(target[last_idx, 11] != 0)) # + int(target[last_idx, 16]!= 0))
             data_ctr[index_found].append(file)
 
@@ -66,28 +122,28 @@ def balance_data(files, dest_path):
                         data_1['mov']['right'].append(file)
                 else:
                     imm_files.append(file)
-                    if target[last_idx, 10] > 0:
+                    if target[last_idx, 12] > 0:
                         # right side immovable
                         data_1['imm']['left'].append(file)
                     else:
                         # left side immovable
                         data_1['imm']['right'].append(file)
 
-            # if index_found == 2:
-            #     if target[last_idx, 3] > 0:
-            #         if target[last_idx, 10] > 0:
-            #             # right side immovable
-            #             data_2['left']['left'].append(file)
-            #         else:
-            #             # left side immovable
-            #             data_2['left']['right'].append(file)
-            #     else:
-            #         if target[last_idx, 10] > 0:
-            #             # right side immovable
-            #             data_2['right']['left'].append(file)
-            #         else:
-            #             # left side immovable
-            #             data_2['right']['right'].append(file)
+            if index_found == 2:
+                if target[last_idx, 3] > 0:
+                    if target[last_idx, 12] > 0:
+                        # right side immovable
+                        data_2['left']['left'].append(file)
+                    else:
+                        # left side immovable
+                        data_2['left']['right'].append(file)
+                else:
+                    if target[last_idx, 12] > 0:
+                        # right side immovable
+                        data_2['right']['left'].append(file)
+                    else:
+                        # left side immovable
+                        data_2['right']['right'].append(file)
 
             ctr += 1
         except Exception as e:
@@ -110,8 +166,20 @@ def balance_data(files, dest_path):
     # count = min(len(data_1['mov']['left']), len(data_1['mov']['right']))
 
     # print(len(data_ctr[0]))
+            
+    # print(len(data_ctr[0]))
 
     # print(len(data_1['mov']['left']), len(data_1['mov']['right']), len(data_1['imm']['left']), len(data_1['imm']['right']))
+
+    # print(len(data_2['left']['left']), len(data_2['left']['right']), len(data_2['right']['left']), len(data_2['right']['right']))
+
+    min_count = min(min(len(data_2['left']['left']), len(data_2['left']['right']), len(data_2['right']['left']), len(data_2['right']['right'])), min(len(data_1['mov']['left']), len(data_1['mov']['right']), len(data_1['imm']['left']), len(data_1['imm']['right'])))
+
+    # print(len(data_ctr[0]), min_count)
+
+    # d = random.sample(data_2['left']['left'], min_count) + random.sample(data_2['left']['right'], min_count) + random.sample(data_2['right']['left'], min_count) + random.sample(data_2['right']['right'], min_count) + random.sample(data_1['mov']['left'], min_count) + random.sample(data_1['mov']['right'], min_count) + random.sample(data_1['imm']['left'], min_count) + random.sample(data_1['imm']['right'], min_count)
+
+    # print(len(d), len(data_ctr[0]))
 
 
 
@@ -130,14 +198,18 @@ def balance_data(files, dest_path):
     #     2: balance_data_2,
     # }
 
-    data = {
-        0: data_ctr[0],
-        1: data_ctr[1],
-        2: data_ctr[2],
-        'mv': mv_files,
-        'imm': imm_files,
-    }
+    data_0_consolidated = data_ctr[0]  # random.sample(data_ctr[0], min_count)
+    data_1_consolidated = data_ctr[1] # random.sample(data_1['mov']['left'], min_count) + random.sample(data_1['mov']['right'], min_count) + random.sample(data_1['imm']['left'], min_count) + random.sample(data_1['imm']['right'], min_count)
+    data_2_consolidated = data_ctr[2] # random.sample(data_2['left']['left'], min_count) + random.sample(data_2['left']['right'], min_count) + random.sample(data_2['right']['left'], min_count) + random.sample(data_2['right']['right'], min_count)
 
+    data = {
+        0: data_0_consolidated,
+        1: data_1_consolidated,
+        2: data_2_consolidated
+        # 'mv': mv_files,
+        # 'imm': imm_files,
+    }
+    
     # count = min(len(data_ctr[0]), len(data_ctr[1]), len(data_ctr[2]))
     # data = {
     #     0: random.sample(data_ctr[0], count//2),
@@ -149,7 +221,7 @@ def balance_data(files, dest_path):
         pickle.dump(data, f)
 
 def main(all_files, dest_path):
-    num_workers = 30
+    num_workers = 40
     # all_files = glob.glob(str(data_path/'*/*.npz'))
     print(len(all_files))
 
@@ -187,12 +259,12 @@ if __name__ == '__main__':
 
     root_path = '/common/users/dm1487/legged_manipulation_data_store'
     root_traj_path = f'{root_path}/trajectories'
-    sub_path = 'iros24_play_feb23_same_policy/2_obs'
+    sub_path = 'iros24_play_feb21/2_obs'
 
     main_folder = f'{root_traj_path}/{sub_path}'
     data_folder = f'{main_folder}/all_data'
     tmp_folder = f'{main_folder}/tmp'
-    balance_folder = f'{main_folder}/balanced'
+    balance_folder = f'{main_folder}/no_balanced'
 
 
     # save_id = 2 -> has only 0 and 1 obs data
@@ -221,10 +293,9 @@ if __name__ == '__main__':
             all_files[0] += data[0]
             all_files[1] += data[1]
             all_files[2] += data[2]
-            all_files['mv'] += data['mv']
-            all_files['imm'] += data['imm']
+            # all_files['mv'] += data['mv']
+            # all_files['imm'] += data['imm']
     
-    # min_len = min(len(all_files['mv']), len(all_files['imm']))
     data_len = min(len(all_files[0]), len(all_files[1]), len(all_files[2]))
 
     print(data_len)
@@ -236,11 +307,13 @@ if __name__ == '__main__':
     
     for key in all_files.keys():
         with open(f'{balance_folder}/train_{key}.pkl', 'wb') as f:
-            if key == 'mv' or key == 'imm':
-                pickle.dump(all_files[key][:data_len], f)
-            else:
-                pickle.dump(all_files[key][:data_len], f)
+            # if key == 'mv' or key == 'imm':
+            #     random.shuffle(all_files[key])
+            #     pickle.dump(all_files[key][:data_len], f)
+            # else:
+            pickle.dump(all_files[key], f)
 
-    shutil.rmtree(tmp_folder)
+    if os.path.exists(tmp_folder):
+        shutil.rmtree(tmp_folder)
